Move predict to base angle dense head
